zshrc: Introduce vcs_info()
This is supposed to be a better __vcs_dir(). In zsh, functions starting
with an underscore are supposed to be part of the completion system.
Therefore, the name switch to vcs_info().

This commit does not enable vcs_info(), yet.

It currently supports git, hg, bzr, svk and svn. However, only the
git, hg and svn backends were tested. Up to now, only git supports
*all* functionality vcs_info() is able to provide.

Its configuration is documented in a comment within the code.