zshrc: Improve robustness of the bzr backend in vcs_info()
I just played a little with bzr; calling 'bzr' itself in the prompt is
painfully slow. And we do it *twice* for the base directory and the
branchname/revision-number.

I don't know enough about bzr, to change this.
Personally, I would replace the branch-root, by the base directory, that
VCS_INFO_detect_by_dir().

The branch name *could* be replaced by ${bzrbase:t}, but since it is
retrieved from the same command as the revision-number, so that will not
save us a call to 'bzr'...

Hints on this one are highly appreciated. Like this, the bzr backend
is just too slow to use (for me on my laptop at 1.5GHz).