zshrc: Add experimental bzr backend to vcs_info()
I works without calling 'bzr' itself, but remains disabled by default,
because I have *no* idea if it does the right thing for all bzr repositories.

If a bzr user complains about the slowness of the default bzr backend, we can
now point him/her to:
    % zstyle ':vcs_info:bzr' use-simple true

And get feedback, if that works in all cases.