zshrc: Fixing $PS2
We use prompt_subst for vcs_info() to work. PS2 contained a backtick, which
starts off command substitution. There was no closing backtick (because the
backtick is meant to be there for cosmetic reasons only), so that threw
a parsing error now. Thanks to Matthias Merk noticing.

This also removes some dead comments.