zshrc: provide .. as a completion
There was a bug in some 4.3.x versions of zsh, that enabled the
completion of .. by default. That was a bug!

This change does *not* provide '.' as a possible completion. But that
is easily possible and an according snippet is included among the
examples in etc/skel/.zshrc (commented out, as usual).