adding -f (force) switch to xtrename
xtrename checks if $DISPLAY is set before printing the escape sequence that
changes the xterm title. However, in certain circumstances this may not be an
optimal check. For example: if the shell is running on a remote server which is
headless, but an xterm is being used locally. In this case you can now use
the -f switch to force the escape sequence to be issued regardless of the state
of $DISPLAY.