zshrc: add wrapper for vim
This wrapper sets VIM_PLEASE_SET_TITLE, so the vimrc knows to set screens
hardstatus.

The wrapper also supports an array ($VIM_OPTIONS), that can be used to set
options for every vim invocation.

To use the installed vim binary without the wrapper function, use either of these:
  % command vim
  % =vim

The latter needs the EQUALS option set in zsh (which is the default).