zsh: move completion.d and functions.d to usr/share
Having an additional set of files in /etc sucks because there is no
reliable way to deal with removing files and user customizations in any
sane way.

Keeping completions and additional function files in usr/share makes a
lot more sense. This also gives us the opportunity to finally get rid of
one of the annoying warning-tests at the start of our zshrc.