zshrc: Insert a condition for loading /etc/default/locale
In 76706fcdd9c160d7f3212e597d2139f45531da72 we started loading
/etc/default/locale unconditionally in order to pick up sane defaults
for a given system.

While this is a good idea in most cases, because users may still
customize the locale for their interactive sessions in ~/.zshrc.local,
this does cause problems when you need most of your shells in the
default locale and a few in another from time to time.

That is the case for 'iso-term' a script that changes utf8 parts of the
default environment to iso885915 and then calls an x11 terminal
emulator.

This change enables us to conditionally turn off sourcing the default
locale file by inserting ZSH_NO_DEFAULT_LOCALE=1 into the environment of
the calling process.