zsh: Import new version of mercurial completion _hg
We're shipping a completion for the mercurial VCS since quite some time.
This version was recently submitted to the zsh-workers mailing list with
a new licence so that it can be distributed along with zsh.

We can drop this completion once the zsh version that includes an _hg
completer is widely used (or at least included in grml).