zshrc: actively do: "setopt unset"
It seems like some people think it is a good idea to unset that option.
That is a very bad idea though.

Why? Well, because of this:
% unset foo
% echo "foo$foo"
zsh: foo: parameter not set

That is totally incompatible with every other bourne like shell out
there. And it breaks many shell idioms.

In fact, it will break at least 'vcs_info()' and 'accept_line()' for us.
Which is unacceptable.

I also added 'emulate -L zsh' to accept-line() and vcs_info(), which are
two functions, that will fail horribly with 'unset' unset. That will
leave the shell intact even if someone wants to 'setopt no_unset' in
his/her .zshrc.local