zshrc: Switch `minimal-shell()' from ksh93 to mksh
As announced in:
    <http://ml.grml.org/pipermail/grml-devel/2010-February/000011.html>

The minimal-shell() function now checks if the shell in question (mksh)
is available and if so, executes it with an empty environment, with only
TERM and HOME set. The mksh setup used sets a number of other
environment variables (such as $VISUAL, $EDITOR and $PAGER).

One added feature is that the mksh prompt shows non-zero return values
from programs you ran:

[snip]
    (3)-~$ false
    [1]-(4)-~$ true
    (5)-~$
[snap]

The prompt does *not* feature any eye-candy such as standout or colour
sequences. It should therefore run in pretty much every terminal.

Special keys are set up via terminfo if `infocmp' is available.
Otherwise mksh's rather sane defaults are used.