zshrc: Make minimal-shellrc printf-free
One of the major reasons for `minimal-shell()' is to be able to use it
without /usr being mounted. But some OS vendors think it's smart to put
`printf' into /usr/bin - sed, of course happily lives in /bin. Debian
happens to think that's funny, too.

mksh doesn't make `printf' a builtin because it doesn't want to link
libm in (well, you can hack it in - but that's a discouraged non-
default). This replaces "printf '%s' foo' with "print -r -- foo", which
should resolve the issue for us nicely.