Remove startx and xinit functions
Current X11 versions as well as the current grml-x version do
not longer need a configuration file. Therefore the xinit/startx
functions are not longer needed.

The decision to switch to a config less xorg environment was made at the
grml developer meeting http://grml.org/reports/devmeeting_2010/ and a the
new grml-x packet is already uploaded to grml-testing.