zshrc: Do not set `localoptions' in `xsource()'
"emulate -L zsh" sets zsh's option set to the default values of zsh's
native mode. The `-L' causes the options to be set local to the
surrounding function. That makes sense in most cases.

With `xsource()' however, we explicitly want to *allow* changing options
from the files sourced by that function. That's one of the main purposes
of the function.

I took the `emulate' call out entirely because it doesn't make sense
without `-L' in untility functions. I also hardened the function's code
a little when it's run with non-zsh-mode options (like sh_word_split).

Reported-by: Dominic Hopf