Fix `VCS_INFO_realpath()' for zsh 4.3.10 and below
If the chpwd hook is activated by using the `chpwd_functions[]' array,
could break `VCS_INFO_realpath()', which could lead to endless loops in
`vcs_info' if the one of the hook functions produced output.

This fixes it.

I'm refraining from patching the grmlzshrc-included
`VCS_INFO_realpath()' function, because the use of "cd -q" requires a
pretty recent version of the shell.