Fixing ssl-cert-{sha512,sha256,sha1,md5}()
These functions are generated in a loop. However, the way they were
generated was broken. The loop variable is *not* replaced by a literal
string in the resulting functions. Therefore, as soon as the variable
is unset all functions become broken. Luckily, they broke loudly.
Otherwise this may not have been caught easily.

Reported-by: Michael Gebetsroither <gebi@grml.org>
Signed-off-by: Frank Terbeck <ft@grml.org>