Workaround incompatibilities with the keyboard setup of some vendors
Some vendors, like Debian (and therefore Ubuntu), enable smkx terminal
mode in the zle-line-init widget and turn it back off in
zle-line-finish.

While that is the right thing to do, because it makes sure the values
from $terminfo are valid. However, we are currently incompatible with
that. And it is a lot of work to get it right for every binding we have,
because we do have some, that are not available in the vast majority of
terminfo entries (like <Ctrl>-<CursorKey>).

Signed-off-by: Frank Terbeck <ft@grml.org>