zshrc: New prompt: Only enable prompt themes with zsh >= 4.3.7
The themes do use modern features, that require this constraint.

The fallback prompt, that was being used when promptsys wasn't available
is _way_ more backwards compatible (and will remain like that). Use that
fallback in cases of $ZSH_VERSION < 4.3.7 as well.

This also means, that the GRMLPROMPT=1 and [[ -n $GRML_CHROOT ]] prompts
are not available with shells that old. Version 4.3.7 of zsh was
released on 2008-10-28. I think it's reasonable to require ~5 year old
software for the full-blown version of a setup. The last stable debian
release ships version 4.3.10 and the upcoming release will sport version
4.3.17. Grml ISOs come with version >= 5.0.0.

Signed-off-by: Frank Terbeck <ft@grml.org>