Bind insert-datestamp to "ESC d" (Alt-d on many terminals) instead of "Ctrl-e d"
The github issue #62 for the grml-etc-core repository noted a delay with the
default Ctrl-e binding in emacs mode. This is caused by a binding where Ctrl-e
is a prefix. The shells waits for a timeout to figure out if the user wanted
the "Ctrl-e" binding, or or if there's a d coming to complete the "Ctrl-e d"
binding.

Closes: #62