vimrc: include defaults.vim into our configuration
We want to take over most of the configuration options defined in
$VIMRUNTIME/defaults.vim (which is loaded instead the user vimrc
(only) if the user vimrc file does not exist / is not found).

To prevent Vim from loading the defaults.vim again (if the user vimrc
does not exists) we have to set `let g:skip_defaults_vim=1`.

So we include everything of defaults.vim in our configuration and
override only single options we do not want.
(Currently only `set mouse=`).

JFTR, in another step we could just source the defaults.vim file and
override the settings we want to override.

This change is built on the proposal from @mika in the branch https://github.com/grml/grml-etc-core/compare/mika/vimrc

The background discussion can be found here:

https://github.com/grml/grml-etc-core/commit/8cdc0df085211200682699b53e600fcccfe73989#diff-12b5039d82248e5798b382628a15fc63