zshrc: no longer set histignorealldups
The histignorealldups option provides a feature, that if a new command
line being added to the history list duplicates an older one, the older
command is removed from the list.

Demonstration:

| root@grml ~ # echo foo
| foo
| root@grml ~ # echo bar
| bar
| root@grml ~ # history -E 0
|     1  9.7.2021 07:48  exec zsh
|     2  9.7.2021 07:48  echo foo
|     3  9.7.2021 07:48  echo bar
| root@grml ~ # echo foo
| foo
| root@grml ~ # history -E 0
|     1  9.7.2021 07:48  exec zsh
|     3  9.7.2021 07:48  echo bar
|     5  9.7.2021 07:48  echo foo

This is certainly useful for several people, like in personal setups.
But especially for system administration tasks it's quite essential,
that the history shows "the full truth", and it's hard to reconstruct
the order of commands that have been executed.

Whoever wants to get the old behavior back, can enable `setopt
histignorealldups` in their personal zshrc.  Provide the according
configuration snippet in /etc/skel/.zshrc instead, and update docs
accordingly.

Closes: grml/grml-etc-core#114