zshrc: discard stderr from vcs_info
Changing directory to a bare git repository results in error messages
being printed:

fatal: this operation must be run in a work tree

This is really annoying, especially when working with git - you do not
know if your command complained or zsh prompt generated it.

So let's just discard stderr from vcs_info.