Re-introdude ip-screen
In 8bc8f27 we decided to drop ip-screen in favor of hostname -I + sed,
and put the command pair in our screenrc in grml/grml-etc-core to make
grml-scripts-core fit for ARM, see: grml/grml-scripts-core#3.

In grml/grml-etc-core#162 I tried to replace ip-screen in our screenrc
with the hostname -I + sed command pair, but it turned out that to use
it in the backtick command, we need to use "sh -c" and escape "^" with
"\^".

So in the end we decided to re-introdue ip-screen as a wrapper script
for hostname + sed.

Updated the copyright date too.

This partly reverts commit 8bc8f2770edf9c389f5bf3c2b5e0493ff1340d21.

Closes: grml/grml-etc-core#162