zshrc: deprecate xunfunction
It was introduced in commit 3d6452fdfc44310b5615eaf61c95cab4c27d9bf8 in 2007,
and 724508d30c788da653579d44ac3566f27ffed1d6 in 2008 tried to remove it.
The remove was reverted in 347c7def8583733dd8f8440214af25fbaf89666f in 2008.

Lets try this again, this time with a deprecation warning.